+++ CHANGELOG.md
@@ -37,3 +37,11 @@
 ## 0.1.1-nullsafety
 
 * Fixed IOS runtime problem
+
+## 0.1.2-nullsafety
+
+* Fixed android runtime problem
+
+## 0.1.3-nullsafety
+
+* Added support for huawei, honor, poco, xiaomi, redmi, letv phones